driver: Split 'ne2k_pci' to common & pci parts
